Update bin console and setup plus related docs (#1)
Though its written in the README I assumed when it said "run `bin/console`" it meant `. bin/console` or `bundle exec bin/console` and so struggled to get it to work.

Once I realised my foolishness, it then was not loading the project as expected.

So this change covers updating te docs, more as a reminder to myself than a warning to others. Plus updating the 2 files in `bin/` to ensure they actually run and load the project ready for working.