Add fully styled root page (#3)
Rather than just return some text, this change adds the first proper page to the demo site.

This means its adding all the assets needed to properly present the page. It should be noted that the code is essentially copied and pasted from https://github.com/DEFRA/quke-example/tree/master/demo_app.

The goal is to replace the code currently in the example, with a simple reference to this gem thereby making the example cleaner and more true to life.