Move Runner to DemoApp and fix exe
It didn't seem right having the `Runner` for the demo app in the root `Quke` module so this change moves it to the `DemoApp` module.

Also we found that running `exe/quke_demo_app` was throwing an error. A bit of digging and it seems it

- didn't like our require of Quke, particularly when run from [quke-example](https://github.com/DEFRA/quke-example)
- would error complaining it didn't know what `DemoApp` was

So we also ensured that we could start the app directly using `exe/quke_demo_app` and from **quke-exemple**.