Add CLI support to the project
A nice feature to add (plus it gave us some exposure to a couple of new cool tools) was to allow users of the demo app to change the port used by Sinatra. It defaults to 4567, and it's unlikely someone would already be running something on that already.

But just in case we've added the ability to specify a custom port in the app. We've done this by adding [Thor](https://github.com/erikhuda/thor) as a dependency. Thor is a toolkit for building powerful command-line interfaces. In our case it allowed us to

- add a default command
- add a description for that command
- add help on how to use the command
- add support for a custom argument (`--port`)
- add an alias for the custom argument (`-p`)
- add help on how to the custom argument

This meant some refactoring in the project, mainly dropping the `Run` class in favour of a `CLI` class that inherits from Thor.

The second cool tool we added was [Aruba](https://github.com/cucumber/aruba). Aruba allows you to test command-line applications with Cucumber-Ruby, RSpec or Minitest. It essentially runs your app in a separate process but still allows you to make assertions against the behaviour of it.

The one issue we faced was that because it runs things in a separate process, [Simplecov](https://github.com/colszowka/simplecov) was not seeing that certain lines did have test coverage. So a final change was to resolve this based a stackoverflow post we found (https://stackoverflow.com/a/20505441).

So TL;DR

- the app has a proper CLI
- the app has tests for its CLI