Update Travis config for deployment
Using the travis CLI we have updated the `.travis.yml` file in the hopes that we can deploy to <https://rubygems.org> when we push a new tag to the repo.

This means we can better automate the process, rather than relying on a dev to manually push the gem after tagging the version.