Refactor the namespaces
We hit a problem when trying to integrate this with [Quke](https://github.com/DEFRA/quke) in the [Quke example project](https://github.com/DEFRA/quke-example).

We clearly haven't got the namespace right for `Quke` because we found that depending on which order the gems are referenced in the `Gemfile`, one gem would overwrite the other at runtime.

So if **quke** was referenced before **quke-demo-app**, it would bomb when we tried to run `bundle exec quke`.

```bash
bundler: failed to load command: quke (/home/travis/build/DEFRA/quke-example/vendor/bundle/ruby/2.4.0/bin/quke)
NameError: uninitialized constant Quke::Quke
Did you mean?  Quke
  /home/travis/build/DEFRA/quke-example/vendor/bundle/ruby/2.4.0/bundler/gems/quke-611c9adc6410/exe/quke:9:in `<top (required)>'
  /home/travis/build/DEFRA/quke-example/vendor/bundle/ruby/2.4.0/bin/quke:23:in `load'
  /home/travis/build/DEFRA/quke-example/vendor/bundle/ruby/2.4.0/bin/quke:23:in `<top (required)>'
```

Not that it mattered so much, but if you listed them the overway round, any calls to anything in **quke-example-project** would give the same result.

We didn't want to leave the solution as ensure the dependencies are loaded in a certain order. But also this is supposed to be a simple little app just used as an example.

So rather than invest in trying to get to the bottom of why the 2 are incompatible, we've opted to simplify the namespace in this project.