Switch to SonarCloud from CodeClimate
https://sonarcloud.io/organizations/defra/projects

We used to use [Code Climate](https://codeclimate.com/) for our code quality checks and test coverage analysis. Then a number of Defra projects started using an internal SonarQube instance, and some public ones used SonarCloud (the cloud SASS version of SonarQube).

We've since adopted SonarCloud formerly and all projects will coalesce onto. This is the last of our repos to switch across to SonarCloud.

** Notes

The figuring out of how to do this (because it wasn't straight forward!) was done in

- https://github.com/DEFRA/waste-exemptions-front-office/pull/317
- https://github.com/DEFRA/waste-exemptions-front-office/pull/318